Refactored out labels.
